porting changes 291 into ui 73

+++ src/app/shared/comcol/comcol-forms/edit-comcol-page/comcol-role/comcol-role.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Group } from '../../../../../core/eperson/models/group.model';
 import { Community } from '../../../../../core/shared/community.model';
 import { BehaviorSubject, Observable } from 'rxjs';
@@ -72,6 +73,7 @@ export class ComcolRoleComponent implements OnInit {
   roleName$: Observable<string>;
 
   constructor(
+    protected modalService: NgbModal,
     protected requestService: RequestService,
     protected groupService: GroupDataService,
     protected notificationsService: NotificationsService,
@@ -122,22 +124,28 @@ export class ComcolRoleComponent implements OnInit {
 
   /**
    * Delete the group for this community or collection role.
+   * OSPR UI-13 issue 291: Let user to confirm the delete operation.
    */
-  delete() {
-    this.groupService.deleteComcolGroup(this.groupLink).pipe(
-      getFirstCompletedRemoteData()
-    ).subscribe((rd: RemoteData<NoContent>) => {
-      if (rd.hasSucceeded) {
-        this.groupService.clearGroupsRequests();
-        this.requestService.setStaleByHrefSubstring(this.comcolRole.href);
-      } else {
-        this.notificationsService.error(
-          this.roleName$.pipe(
-            switchMap(role => this.translateService.get('comcol-role.edit.delete.error.title', { role }))
-          ),
-          rd.errorMessage
-        );
-      }
+   public confirmDelete(content) {
+    this.modalService.open(content, {backdrop:'static'}).result.then(
+      (result) => {
+        if (result === 'ok') {
+          this.groupService.deleteComcolGroup(this.groupLink).pipe(
+            getFirstCompletedRemoteData()
+          ).subscribe((rd: RemoteData<NoContent>) => {
+            if (rd.hasSucceeded) {
+              this.groupService.clearGroupsRequests();
+              this.requestService.setStaleByHrefSubstring(this.comcolRole.href);
+            } else {
+              this.notificationsService.error(
+                this.roleName$.pipe(
+                  switchMap(role => this.translateService.get('comcol-role.edit.delete.error.title', { role }))
+                ),
+                rd.errorMessage
+              );
+            }
+          });
+        }
     });
   }
